move FetchError to its own file to satisfy linter
