+++ src/common/utils/fetch-client.ts
@@ -1,3 +1,5 @@
+import { FetchError } from './fetch-error';
+
 export class FetchClient {
   constructor(readonly baseURL: string, readonly options?: RequestInit) {}
 
@@ -89,24 +91,6 @@ export class FetchClient {
     }
 
     return response;
-  }
-}
-
-export class FetchError<T> extends Error {
-  readonly name: string = 'FetchError';
-  readonly message: string = 'The request could not be completed.';
-  readonly response: { status: number; headers: Headers; data: T };
-
-  constructor({
-    message,
-    response,
-  }: {
-    message: string;
-    readonly response: FetchError<T>['response'];
-  }) {
-    super(message);
-    this.message = message;
-    this.response = response;
   }
 }
